requirements_runpod.txt
# Core dependencies
diffusers==0.27.2
transformers==4.36.2
accelerate>=0.20.0
safetensors>=0.3.1
huggingface-hub>=0.16.4

# Image processing
opencv-python-headless>=4.8.0
pillow==10.2.0
einops>=0.6.1

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
peft>=0.4.0

# RunPod
runpod>=1.3.0

# Utilities
numpy>=1.24.0
protobuf==3.20
scipy>=1.10.0

# Optional but recommended
xformers>=0.0.20
